requirements.txt
# Pinned dependencies for comfyui-agent
# Generated from: pip install -e .[dev,mcp]
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.14.3
black==25.12.0
blinker==1.9.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.4
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
docutils==0.22.4
exceptiongroup==1.2.2
fastapi==0.128.0
filelock==3.20.0
Flask==3.1.2
frozenlist==1.8.0
google-auth==2.47.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx-sse==0.4.3
httpx==0.28.1
hyperframe==6.1.0
hypothesis==6.151.2
id==1.5.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-specifications==2025.9.1
jsonschema==4.26.0
keyring==25.7.0
lark==1.3.1
librt==0.7.8
lxml==6.0.2
MarkupSafe==2.1.5
mcp==1.26.0
more-itertools==10.8.0
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
nh3==0.3.2
numpy==2.2.6
olefile==0.47
orjson==3.11.7
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic-settings==2.12.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyproject_hooks==1.2.0
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.3.0
pywin32-ctypes==0.2.3
pywin32==311
PyYAML==6.0.3
readme_renderer==44.0
referencing==0.37.0
regex==2025.11.3
requests-toolbelt==1.0.0
requests==2.32.5
rfc3986==2.0.0
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.10
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.45
sse-starlette==3.2.0
starlette==0.50.0
tomli==2.4.0
tomli_w==1.2.0
twine==6.2.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uv==0.9.22
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.5
wheel==0.46.3
yarl==1.22.0
zipp==3.23.0
